requirements.txt
pandas
sqlalchemy
pymysql
pyodbc
cryptography
cloud-sql-python-connector[pymysql]
fastapi
uvicorn
google-auth
pydantic
requests
PyYAML
pydantic-settings
python-multipart
